Refactor: Add toggle switch component to Header
This commit adds a new togle swchgit component to the Header component in order to provide a visual representation of a switch for a specific feature. The switch component is implemented in a separate file called Switch.jsx and is imported and used in the Header component. This enhancement improves the user experience and adds a new interactive element to the application.